enable nativeWebTap setting for iOS
