+++ src/main/java/io/appium/java_client/android/AndroidDriver.java
@@ -47,7 +47,7 @@ import java.net.URL;
 public class AndroidDriver<T extends WebElement>
     extends AppiumDriver<T>
     implements PressesKeyCode, HasNetworkConnection, PushesFiles, StartsActivity,
-        FindsByAndroidUIAutomator<T>, LocksAndroidDevice, HasSettings, HasDeviceDetails,
+        FindsByAndroidUIAutomator<T>, LocksAndroidDevice, HasAndroidSettings, HasDeviceDetails,
             HasSupportedPerformanceDataType {
 
     private static final String ANDROID_PLATFORM = MobilePlatform.ANDROID;
