+++ src/main/java/io/appium/java_client/ios/HasIOSSettings.java
@@ -1,0 +1,32 @@
+/*
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * See the NOTICE file distributed with this work for additional
+ * information regarding copyright ownership.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package io.appium.java_client.ios;
+
+
+import io.appium.java_client.HasSettings;
+import io.appium.java_client.Setting;
+
+interface HasIOSSettings extends HasSettings {
+    /**
+     * Set the `nativeWebTap` setting. *iOS-only method*.
+     * Sets whether Safari/webviews should convert element taps into x/y taps
+     * @param enabled turns nativeWebTap on if true, off if false
+     */
+    default void nativeWebTap(Boolean enabled) {
+        setSetting(Setting.NATIVE_WEB_TAP, enabled);
+    }
+}
